create file for false positive violation removal method
